feat: refactor email backend registration and update documentation for SES support
