+++ pyproject.toml
@@ -44,7 +44,6 @@ smtp = ["aiosmtplib>=3.0.0"]
 ses = ["botocore>=1.31.0"]
 httpx = ["httpx>=0.27.0"]
 aiohttp = ["aiohttp>=3.9.0"]
-all = ["aiosmtplib>=3.0.0", "httpx>=0.27.0", "aiohttp>=3.9.0", "botocore>=1.31.0"]
 
 ######################
 # Build & Versioning #
