+++ src/tests/test_transport_headers.py
@@ -30,7 +30,7 @@ async def test_httpx_transport_post_with_custom_headers() -> None:
         await transport.post("https://api.example.com", json={"key": "value"}, headers=custom_headers)
 
         # Verify headers were passed to post call
-        args, kwargs = mock_client.post.call_args
+        _args, kwargs = mock_client.post.call_args
         assert kwargs.get("headers") == custom_headers
 
 
@@ -56,5 +56,5 @@ async def test_aiohttp_transport_post_with_custom_headers() -> None:
         await transport.post("https://api.example.com", json={"key": "value"}, headers=custom_headers)
 
         # Verify headers were passed to post call
-        args, kwargs = mock_session.post.call_args
+        _args, kwargs = mock_session.post.call_args
         assert kwargs.get("headers") == custom_headers
